add arch test
add integration test and refactor TutorialGroup code
